requirements.txt
Package                 Version
----------------------- -------------
absl-py                 0.9.0
annoy                   1.17.0
apt-clone               0.2.1
apturl                  0.5.2
argon2-cffi             21.1.0
astunparse              1.6.3
attrs                   21.2.0
backcall                0.2.0
beautifulsoup4          4.8.2
bleach                  4.1.0
blinker                 1.4
Brlapi                  0.7.0
cachetools              4.1.1
certifi                 2021.5.30
cffi                    1.14.6
chardet                 3.0.4
charset-normalizer      2.0.4
click                   7.1.2
command-not-found       0.3
configobj               5.0.6
cryptography            2.8
cupshelpers             1.0
cycler                  0.10.0
Cython                  0.29.14
dbus-python             1.2.16
decorator               4.4.2
defer                   1.0.6
defusedxml              0.7.1
distro                  1.4.0
entrypoints             0.3
fasteners               0.16.3
filelock                3.0.12
future                  0.18.2
gast                    0.3.3
gensim                  3.8.3
geomloss                0.2.4
google-auth             1.19.2
google-auth-oauthlib    0.4.1
google-pasta            0.2.0
grpcio                  1.30.0
grpcio-tools            1.30.0
h5py                    3.4.0
httplib2                0.14.0
idna                    3.2
ifaddr                  0.1.6
ipykernel               5.3.4
ipython                 7.17.0
ipython-genutils        0.2.0
jedi                    0.17.2
Jinja2                  3.0.1
joblib                  0.16.0
jsonpatch               1.26
jsonpointer             2.0
jsonschema              3.2.0
jupyter-client          6.1.6
jupyter-core            4.6.3
jupyterlab-pygments     0.1.2
Keras                   2.2.4
Keras-Applications      1.0.6
Keras-Preprocessing     1.1.2
keyring                 18.0.1
kiwisolver              1.0.1
launchpadlib            1.10.13
lazr.restfulclient      0.14.2
lazr.uri                1.0.3
louis                   3.12.0
lxml                    4.5.2
lz4                     3.1.3
macaroonbakery          1.3.1
Mako                    1.1.0
Markdown                3.2.2
MarkupSafe              2.0.1
matplotlib              3.1.2
mistune                 0.8.4
monotonic               1.6
mpi4py                  3.0.3
munch                   2.5.0
nbclient                0.5.4
nbconvert               6.1.0
nbformat                5.1.3
nemo-emblems            4.6.0
nest-asyncio            1.5.1
netifaces               0.10.4
nltk                    3.6.2
nose                    1.3.7
notebook                6.4.3
numpy                   1.21.2
oauthlib                3.1.0
onboard                 1.4.1
opt-einsum              3.3.0
packaging               20.4
PAM                     0.4.2
pandas                  1.0.5
pandocfilters           1.4.3
parameterized           0.7.0
parso                   0.7.1
pexpect                 4.6.0
pickleshare             0.7.5
Pillow                  7.0.0
pip                     21.2.4
plotly                  4.9.0
prometheus-client       0.11.0
prompt-toolkit          3.0.5
protobuf                3.12.2
psutil                  5.5.1
ptyprocess              0.7.0
pyasn1                  0.4.8
pyasn1-modules          0.2.8
pycairo                 1.16.2
pycparser               2.20
pycrypto                2.6.1
pycups                  1.9.73
pycurl                  7.43.0.2
pydot                   1.4.1
Pygments                2.6.1
PyGObject               3.36.0
pygpu                   0.7.6
PyICU                   2.4.2
pyinotify               0.9.6
PyJWT                   1.7.1
pymacaroons             0.13.0
pymagnitude             0.1.143
PyNaCl                  1.3.0
pynvml                  11.0.0
pyparsing               2.4.6
pyparted                3.11.2
pyRFC3339               1.1
pyrsistent              0.18.0
PySocks                 1.6.8
python-apt              2.0.0
python-dateutil         2.8.1
python-debian           0.1.36ubuntu1
python-xapp             2.0.1
python-xlib             0.23
pytorch-crf             0.7.2
pytz                    2019.3
pyxdg                   0.26
PyYAML                  5.3.1
pyzmq                   19.0.1
regex                   2020.7.14
reportlab               3.5.34
requests                2.26.0
requests-oauthlib       1.3.0
requests-unixsocket     0.2.0
retrying                1.3.3
rsa                     4.6
sacremoses              0.0.43
scikit-learn            0.23.1
scipy                   1.4.1
screen-resolution-extra 0.0.0
seaborn                 0.10.1
SecretStorage           2.3.1
Send2Trash              1.8.0
sentencepiece           0.1.91
setproctitle            1.1.10
setuptools              45.2.0
simplejson              3.16.0
six                     1.16.0
sklearn                 0.0
smart-open              2.2.0
soupsieve               1.9.5
ssh-import-id           5.10
systemd-python          234
tensorboard             2.2.2
tensorboard-plugin-wit  1.7.0
tensorflow              2.2.0
tensorflow-estimator    2.2.0
termcolor               1.1.0
terminado               0.12.1
testpath                0.5.0
Theano                  1.0.4
threadpoolctl           2.1.0
tinycss2                1.0.2
tld                     0.9.1
tokenizers              0.8.1rc1
torch                   1.7.1
torchfile               0.1.0
torchnet                0.0.4
torchtext               0.8.1
torchvision             0.8.2
tornado                 6.1
tqdm                    4.62.2
traitlets               5.1.0
transformers            3.0.2
tweet-preprocessor      0.6.0
typing-extensions       3.10.0.2
ubuntu-advantage-tools  20.3
ubuntu-drivers-common   0.0.0
ufw                     0.36
Unidecode               1.1.1
urllib3                 1.26.6
vboxapi                 1.0
visdom                  0.1.8.9
wadllib                 1.3.3
wapiti3                 3.0.3
wcwidth                 0.2.5
webencodings            0.5.1
websocket-client        0.57.0
Werkzeug                1.0.1
wheel                   0.34.2
wrapt                   1.12.1
xkit                    0.0.0
XStatic-smart-table     1.4.13.2
xxhash                  2.0.2
yaswfp                  0.9.3
zeroconf                0.24.4
